Adds css class to trunk8ed element
This enables you to style your trunk8ed text differently than non trunk8ed text. The css class is removed after restoring the text to  original.
This makes it possible to add a cursor:pointer to the text when trunk8ed to signal that the title can be clicked to show the full text.